requirements.txt
requests
python-dotenv
python-dotenv
fastapi
uvicorn
pydantic